outsource csv components from core to new file and rename links
