+++ superset/db_engine_specs/base.py
@@ -418,7 +418,7 @@ class BaseEngineSpec:  # pylint: disable=too-many-public-methods
             "sep": form_data["delimiter"],
             # frontend already does int-check, check again in case of tampering
             "header": int(form_data["headerRow"])
-            or 0,  # 0 if not form_data["headerRow"] else int(form_data["headerRow"]),
+            or 0,
             "index_col": None
             if not form_data["indexColumn"]
             else int(form_data["indexColumn"]),
